Remove a few layers of wrappers
